fix deprecation warnings
Prior to this commit, this module was producing two deprecation
warnings: one from a change in how Puppet processes templates and
another from a module dependency. This commit changes the
`serverhide.erb` template to use the accepted variable format of `@name`
rather than relying upon the now-deprecated process of using `name`
which channels up through Ruby's `method_missing` to eventually find
`@name`. This commit also removes the `include concat::setup` because the
`puppetlabs/concat` module deprecated that behaviour in 1.0.0, and we
already force a dependency of `puppetlabs/concat` of `>= 1.0.0`.